requirements.txt
fastapi
uvicorn
opencv-python<4.13
pillow
numpy>=1.26,<2
pydantic
python-multipart
aiofiles
torch
timm>=1.0.17
tqdm
ftfy==6.1.1
regex
iopath>=0.1.10
typing_extensions
huggingface_hub
einops
pycocotools
psutil
